chore: move services to the related internal/controller/services package (#1856)
